# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "lsproxy",
    "description": "",
    "license": {
      "name": "Apache-2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0"
    },
    "version": "0.1.5"
  },
  "servers": [
    {
      "url": "http://localhost:4444/v1",
      "description": "API server v1"
    }
  ],
  "paths": {
...
    "/symbol/find-references": {
      "post": {
        "tags": [
          "symbol"
        ],
        "summary": "Find all references to a symbol",
        "description": "The input position should point to the identifier of the symbol you want to get the references for.\n\nReturns a list of locations where the symbol at the given position is referenced.\n\nThe returned positions point to the start of the reference identifier.\n\ne.g. for `User` on line 0 of `src/main.py`:\n```\n 0: class User:\n input____^^^^\n 1:     def __init__(self, name, age):\n 2:         self.name = name\n 3:         self.age = age\n 4:\n 5: user = User(\"John\", 30)\n output____^\n```",
        "operationId": "find_references",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GetReferencesRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "References retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReferencesResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/workspace/list-files": {
      "get": {
        "tags": [
          "workspace"
        ],
        "summary": "Get a list of all files in the workspace",
        "description": "Returns an array of file paths for all files in the current workspace.\n\nThis is a convenience endpoint that does not use the underlying Language Servers directly, but it does apply the same filtering.",
        "operationId": "list_files",
        "responses": {
          "200": {
            "description": "Workspace files retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/workspace/read-source-code": {
      "post": {
        "tags": [
          "workspace"
        ],
        "summary": "Read source code from a file in the workspace",
        "description": "Returns the contents of the specified file.",
        "operationId": "read_source_code",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FileRange"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Source code retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadSourceCodeResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "CodeContext": {
        "type": "object",
        "required": [
          "range",
          "source_code"
        ],
        "properties": {
          "range": {
            "$ref": "#/components/schemas/FileRange"
          },
          "source_code": {
            "type": "string"
          }
        }
      },
...
      "FilePosition": {
        "type": "object",
        "description": "Specific position within a file.",
        "required": [
          "path",
          "position"
        ],
        "properties": {
          "path": {
            "type": "string",
            "example": "src/main.py"
          },
          "position": {
            "$ref": "#/components/schemas/Position"
          }
        }
      },
      "FileRange": {
        "type": "object",
        "required": [
          "path",
          "start",
          "end"
        ],
        "properties": {
          "end": {
            "$ref": "#/components/schemas/Position",
            "description": "The end position of the range."
          },
          "path": {
            "type": "string",
            "description": "The path to the file.",
            "example": "src/main.py"
          },
          "start": {
            "$ref": "#/components/schemas/Position",
            "description": "The start position of the range."
          }
        }
      },
...
      "GetReferencesRequest": {
        "type": "object",
        "required": [
          "identifier_position"
        ],
        "properties": {
          "identifier_position": {
            "$ref": "#/components/schemas/FilePosition"
          },
          "include_code_context_lines": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "description": "Whether to include the source code of the symbol in the response.\nDefaults to none.",
            "example": 5,
            "minimum": 0
          },
          "include_raw_response": {
            "type": "boolean",
            "description": "Whether to include the raw response from the langserver in the response.\nDefaults to false.",
            "example": false
          }
        }
      },
...
      "Identifier": {
        "type": "object",
        "required": [
          "name",
          "range"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "range": {
            "$ref": "#/components/schemas/FileRange"
          }
        }
      },
...
      "Position": {
        "type": "object",
        "required": [
          "line",
          "character"
        ],
        "properties": {
          "character": {
            "type": "integer",
            "format": "int32",
            "description": "0-indexed character index.",
            "example": 5,
            "minimum": 0
          },
          "line": {
            "type": "integer",
            "format": "int32",
            "description": "0-indexed line number.",
            "example": 10,
            "minimum": 0
          }
        }
      },
      "ReadSourceCodeResponse": {
        "type": "object",
        "required": [
          "source_code"
        ],
        "properties": {
          "source_code": {
            "type": "string"
          }
        }
      },
...
      "ReferencesResponse": {
        "type": "object",
        "description": "Response to a references request.\n\nPoints to the start position of the symbol's identifier.\n\ne.g. for the references of `User` on line 0 character 6 of `src/main.py` with the code:\n```\n0: class User:\n1:     def __init__(self, name, age):\n2:         self.name = name\n3:         self.age = age\n4:\n5: user = User(\"John\", 30)\n_________^\n6:\n7: print(user.name)\n```\nThe references will be `[{\"path\": \"src/main.py\", \"line\": 5, \"character\": 7}]`.",
        "required": [
          "references",
          "selected_identifier"
        ],
        "properties": {
          "context": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "$ref": "#/components/schemas/CodeContext"
            },
            "description": "The source code around the references."
          },
          "raw_response": {
            "description": "The raw response from the langserver.\n\nhttps://microsoft.github.io/language-server-protocol/specifications/lsp/3.17/specification/#textDocument_references"
          },
          "references": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FilePosition"
            }
          },
          "selected_identifier": {
            "$ref": "#/components/schemas/Identifier",
            "description": "The identifier that was \"clicked-on\" to get the references."
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearer_auth": []
    }
  ],
...
}